Consolidate to Episode and EpisodeListItem
When getting episode information from the database/api, it will always be deserialized into a pod_req::Episode struct. The Episode struct contains all possible fields so it is compatible. This eliminates EpisodeTrait and the need to Box various types of Episodes.

Certain fields in the Episode have been changed from Option<T> to simply T. For example, with listen_duration a None is effectively the same value as 0. This eliminated needless unwrapping when None and the default value have no useful distinction.

This also uses EpisodeListItem to render episodes in the web ui. This still has a few bugs to work out at the moment, but the majority is working as expected. This puts all logic required to render an EpisodeListItem in episodelistitem.rs to eliminate redundant code.

Various little things throughout have been tidied up as well.