Fix Episode page
Episode page can take a url param `episode_id` that points to an episode in the db to pull metadata from. If this is not present, the method using other params (podcast_title, etc) are used instead. The logic for getting the metadata is mostly unchanged. The global AppState is used less, and Episode pages (with appropriate url params) can be visited directly through the url.

Page state is managed in an enum that allows for Loading, Error, and Ok states.

Some field name discrepencies in Episode are corrected with serde aliases.